Pdf_speech_app/requirements.txt
PyPDF2
tiktoken
Streamlit
langchain
openai
python-docx
